+++ apps/repos/views.py
@@ -23,6 +23,7 @@ from apps.repos.services import (
     repository_performance_summary,
     repository_search_queryset,
     similar_repositories_for_repository,
+    visible_repository_queryset,
 )
 
 AWESOME_LIST_SCAN_TASK_GROUP = "Scan awesome list"
@@ -111,13 +112,14 @@ class RepositorySearchView(ListView):
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
+        visible_repositories = visible_repository_queryset()
         context["awesome_lists"] = (
             AwesomeList.objects.filter(is_active=True)
             .annotate(repo_count=Count("items"))
             .order_by("name")
         )
         context["languages"] = (
-            Repository.objects.exclude(language="")
+            visible_repositories.exclude(language="")
             .values_list("language", flat=True)
             .distinct()
             .order_by("language")
@@ -127,7 +129,7 @@ class RepositorySearchView(ListView):
         params = self.request.GET.copy()
         params.pop("page", None)
         context["querystring"] = params.urlencode()
-        context["total_repositories"] = Repository.objects.count()
+        context["total_repositories"] = visible_repositories.count()
         context["total_lists"] = AwesomeList.objects.filter(is_active=True).count()
         return context
 
@@ -217,7 +219,10 @@ class AwesomeListDetailView(DetailView):
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
         repos = awesome_list_repository_queryset(self.object, self.request.GET)
-        all_list_repos = Repository.objects.filter(awesome_items__awesome_list=self.object)
+        all_list_repos = visible_repository_queryset().filter(
+            awesome_items__awesome_list=self.object
+        )
+        self.object.indexed_repo_count = all_list_repos.count()
         params = self.request.GET.copy()
         params.pop("page", None)
         filter_names = (